cleaned up, moved some files, removed some files, clean slate for tomorrow
