+++ front-end/package.json
@@ -13,7 +13,6 @@
   },
   "dependencies": {
     "jest-cli": "^29.7.0",
-    "pinia": "^3.0.1",
     "vue": "^3.5.13",
     "vue-router": "^4.5.0"
   },
