--- front-end/src/store/folderView.js
@@ -1,10 +1,0 @@
-/*
- *
-* TODO:
-* 1. Setup a store for the main browser view to view folder-contents.
-* *
-* 2. Perhaps set up a store for:
-*   - toggleable settings.
-*   - current open folders (in the tree)
-*
-*/
