+++ front-end/src/tests/bookies/bookies.test.js
@@ -7,6 +7,7 @@ beforeEach(() => {
   c = new Constants;
 })
 
+
 describe("[FlatParent] Rebuilds a folder-item into its original shape.", () => {
   test("Rebuilding regular flat parent.", () => {
     const rebuilt = new FlatParent(c.flatParent, c.flatChildren).parent;
@@ -25,17 +26,19 @@ describe("[FlatParent] Rebuilds a folder-item into its original shape.", () => {
   });
 });
 
-describe("[Rebuild] Rebuilds a Bookies bookmarks array to its original shape.", () => {
-  test("Rebuilding from root of scrambled flat Bookies array.", () => {
-    const rebuilt = new Rebuild(c.scrambledFlatBookies).inflated;
-    expect(rebuilt).toEqual(c.bookiesInflated);
-  })
 
+describe("[Rebuild] Rebuilds a Bookies bookmarks array to its original shape.", () => {
   test("Rebuilding from root of flat Bookies array.", () => {
     const rebuilt = new Rebuild(c.flatBookies).inflated;
     expect(rebuilt).toEqual(c.bookiesInflated);
   })
+
+  test("Rebuilding from root of scrambled flat Bookies array.", () => {
+    const rebuilt = new Rebuild(c.scrambledFlatBookies).inflated;
+    expect(rebuilt).toEqual(c.bookiesInflated);
+  })
 });
+
 
 describe("[Big Rebuild] Rebuilds a big Bookies bookmarks array to its original shape.", () => {
   test("Rebuilding from root of flat Bookies array.", () => {
