+++ front-end/src/views/BrowserView.vue
@@ -1,6 +1,5 @@
 <script setup>
 import FolderTree from '@/components/browser/FolderTree.vue';
-//import FolderTree from '@/components/browser/FolderTreeFlat.vue';
 </script>
 
 <template>
